+++ CHANGELOG.md
@@ -10,3 +10,4 @@
 - MVP-53 - Add login via linked and google
 - MVP-80 - Implement design for modals
 - MVP-174 - Add "welcome" notification
+- MVP-125 - Store user's location